+++ llmmaster/text_to_image_models.py
@@ -1,8 +1,12 @@
+import fal_client
 import requests
+from fal_client import InProgress
 from openai import OpenAI
 from openai.types.images_response import ImagesResponse
 
 from .base_model import BaseModel
+from .config import FLUX1_FAL_TTI_MODELS
+from .config import FLUX1_FAL_TTI_ASPECT_RATIO_LIST
 from .config import MAX_SEED
 from .config import OPENAI_TTI_DEFAULT_N
 from .config import OPENAI_TTI_QUALITY_LIST
@@ -15,6 +19,102 @@ from .config import STABLE_DIFFUSION_TTI_OUTPUT_FORMATS
 from .config import STABLE_DIFFUSION_TTI_STYLE_PRESET_LIST
 from .config import STABLE_DIFFUSION_VERSION
 from .config import REQUEST_OK
+from .config import WAIT_FOR_FLUX1_FAL_TTI_RESULT
+
+
+class Flux1FalTextToImage(BaseModel):
+    '''
+    List of available models as of 2024-09-18
+      - fal-ai/flux/dev
+      - fal-ai/flux/schnell
+    Important:
+    Fal does not support API key import from instance creation.
+    API key must be set in console with `FAL_KEY` environment variable.
+    '''
+    def __init__(self, **kwargs):
+
+        try:
+            super().__init__(**kwargs)
+
+        except Exception as e:
+            msg = 'Error while verifying specific parameters for '
+            msg += 'Flux1FalTextToImage'
+            raise Exception(msg) from e
+
+    def run(self):
+        '''
+        Note:
+        Flux1FalTextToImage returns a dictionary including image url.
+        But when failed to generate, return value is given in str.
+        Handle return value `answer` with care for different type
+        in case of success and failure.
+        '''
+        answer = 'Valid image not generated. '
+
+        try:
+            handler = fal_client.submit(self.parameters['model'],
+                                        arguments=self.parameters['arguments'])
+            flg = True
+            while flg:
+                self._wait(WAIT_FOR_FLUX1_FAL_TTI_RESULT)
+                status = handler.status(with_logs=True)
+                flg = False if not isinstance(status, InProgress) else True
+            answer = handler.get()
+
+        except Exception as e:
+            answer += str(e)
+
+        self.response = answer
+
+    def _verify_arguments(self, **kwargs):
+        '''
+        Expected arguments:
+          - model
+          - prompt
+          - image_size
+          - num_inference_steps
+          - seed
+          - guidance_scale
+          - sync_mode
+          - num_images
+          - enable_safety_checker
+        '''
+        parameters = kwargs
+
+        if kwargs['model'] not in FLUX1_FAL_TTI_MODELS:
+            parameters['model'] = FLUX1_FAL_TTI_MODELS[0]
+
+        arguments = {'prompt': kwargs['prompt']}
+
+        if ('image_size' in kwargs and
+           kwargs['image_size'] in FLUX1_FAL_TTI_ASPECT_RATIO_LIST):
+            arguments['image_size'] = kwargs['image_size']
+
+        if ('num_inference_steps' in kwargs and
+           isinstance(kwargs['num_inference_steps'], int)):
+            arguments['num_inference_steps'] = kwargs['num_inference_steps']
+
+        if 'seed' in kwargs and isinstance(kwargs['seed'], int):
+            arguments['seed'] = kwargs['seed']
+
+        if ('guidance_scale' in kwargs and
+           isinstance(kwargs['guidance_scale'], float)):
+            arguments['guidance_scale'] = kwargs['guidance_scale']
+
+        if 'sync_mode' in kwargs and isinstance(kwargs['sync_mode'], bool):
+            arguments['sync_mode'] = kwargs['sync_mode']
+
+        if 'num_images' in kwargs and isinstance(kwargs['num_images'], int):
+            arguments['num_images'] = kwargs['num_images']
+
+        if ('enable_safety_checker' in kwargs and
+           isinstance(kwargs['enable_safety_checker'], bool)):
+            arguments['enable_safety_checker'] = \
+                kwargs['enable_safety_checker']
+
+        parameters.update(arguments=arguments)
+
+        return parameters
 
 
 class OpenAITextToImage(BaseModel):