+++ tests/test_lumaai_models.py
@@ -14,7 +14,6 @@ from llmmaster import LLMMaster
 
 
 API_KEY = '''
-
 '''